Refactor the Backend from a struct into a trait
Thanks to Apple's new announcement (namely the deprecation of OpenGL on
the Mac) OpenGL can no longer be assumed as a universally supported
graphics API. If Quicksilver wants to support whatever new line of Macs
that ship without OpenGL, it seems prudent for the Backend to be highly
modular and only make assumptions about the architecture of Quicksilver
itself, rather than the underlying graphics API.

This change should not impact performance, as it is fundamentally the
same code at compile time.
